refactor: extract database functions to a testutils package
The reason for this change is that otherwise these functions would have
to be rewritten over and over again for each package, when needing a database
connection.

They are placed on the "internal" package because we don't want them to
be exported.
