+++ main_test.go
@@ -6,14 +6,12 @@ import (
 	"os"
 	"testing"
 
-	"github.com/RedHatInsights/sources-api-go/config"
 	"github.com/RedHatInsights/sources-api-go/dao"
+	"github.com/RedHatInsights/sources-api-go/internal/testutils"
 	l "github.com/RedHatInsights/sources-api-go/logger"
 	m "github.com/RedHatInsights/sources-api-go/model"
 	"github.com/RedHatInsights/sources-api-go/util"
 	"github.com/labstack/echo/v4"
-	"gorm.io/driver/postgres"
-	"gorm.io/gorm"
 )
 
 var (
@@ -38,7 +36,7 @@ func TestMain(t *testing.M) {
 
 	if *createdb {
 		fmt.Fprintf(os.Stderr, "creating database %v...", testDbName)
-		err := createTestDB()
+		err := testutils.CreateTestDB()
 		if err != nil {
 			fmt.Fprintf(os.Stderr, "Error creating test DB: %v", err)
 			os.Exit(1)
@@ -46,7 +44,7 @@ func TestMain(t *testing.M) {
 
 		os.Exit(0)
 	} else if *integration {
-		connectToTestDB()
+		testutils.ConnectToTestDB()
 		getSourceDao = getSourceDaoWithTenant
 		getApplicationDao = getApplicationDaoWithTenant
 		getEndpointDao = getEndpointDaoWithTenant
@@ -86,67 +84,10 @@ func TestMain(t *testing.M) {
 	code := t.Run()
 
 	if *integration {
-		dao.DB.Exec("DROP TABLE endpoints")
-		dao.DB.Exec("DROP TABLE meta_data")
-		dao.DB.Exec("DROP TABLE applications")
-		dao.DB.Exec("DROP TABLE application_types")
-		dao.DB.Exec("DROP TABLE sources")
-		dao.DB.Exec("DROP TABLE source_types")
-		dao.DB.Exec("DROP TABLE tenants")
+		testutils.DropSchema()
 	}
 
 	os.Exit(code)
-}
-
-func connectToTestDB() {
-	db, err := gorm.Open(postgres.Open(testDbString(testDbName)), &gorm.Config{})
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "db must not exist - create the database '%v' first with '-createdb'.", testDbName)
-		panic(err)
-	}
-	dao.DB = db
-	rawDB, err := db.DB()
-	if err != nil {
-		panic(err)
-	}
-	rawDB.SetMaxOpenConns(20)
-
-	// migrate all of the models.
-	err = db.AutoMigrate(
-		&m.SourceType{},
-		&m.ApplicationType{},
-
-		&m.Source{},
-		&m.Application{},
-
-		&m.Endpoint{},
-		&m.MetaData{},
-	)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error automigrating the schema: %v", err)
-		os.Exit(1)
-	}
-}
-
-func createTestDB() error {
-	db, err := gorm.Open(postgres.Open(testDbString("postgres")), &gorm.Config{})
-	if err != nil {
-		panic(err)
-	}
-
-	out := db.Exec(fmt.Sprintf("CREATE DATABASE %v", testDbName))
-	return out.Error
-}
-
-func testDbString(dbname string) string {
-	return fmt.Sprintf(
-		"user=%s password=%s dbname=%s host=%s port=%d sslmode=disable",
-		config.Get().DatabaseUser,
-		config.Get().DatabasePassword,
-		dbname,
-		config.Get().DatabaseHost,
-		config.Get().DatabasePort,
-	)
 }
 
 func AssertLinks(t *testing.T, path string, links util.Links, limit int, offset int) {
